makecttemplate.py: changing to new concept
The new concept: treat each variable in a specification as a "block."  A
block can draw multiple entries from its JSON input object.  A block can
also take up multiple cells in the output table.

What exactly a block draws from the input object, and how it is exactly
displayed in cells, is specified in the "block template."